+++ radix-engine/src/model/substates/global.rs
@@ -18,14 +18,3 @@ impl GlobalAddressSubstate {
         }
     }
 }
-
-#[derive(Debug)]
-pub struct GlobalRENode {
-    pub address: GlobalAddressSubstate,
-}
-
-impl GlobalRENode {
-    pub fn node_deref(&self) -> RENodeId {
-        self.address.node_deref()
-    }
-}
